Redesign of database
The new database design allows documents to be accessible only via a URI that includes a date instead of via  just the slug.

This allows slugs to be non-unique.

The permanent table lets pages be promoted to be accessible without a date (Perl to handle this isn't implemented yet).

This makes the system a bit more like Wordpress. Documents with an entry in the Permanent table are akin to Wordpress' "Pages".